@@ -65,7 +65,9 @@ class ExtensionInstallStorage extends InstallStorage {
     return new static(
       $this->configStorage,
       $this->directory,
-      $collection
+      $collection,
+      $this->includeProfile,
+      $this->installProfile
     );
   }
 
