@@ -4,6 +4,8 @@ namespace Drupal\Tests\media\Kernel;
 
 use Drupal\KernelTests\KernelTestBase;
 use Drupal\media\Entity\Media;
+use Drupal\media\OEmbed\Provider;
+use Drupal\media\OEmbed\ResourceFetcher;
 use Drupal\media\OEmbed\UrlResolverInterface;
 use Drupal\media\Plugin\Validation\Constraint\OEmbedResourceConstraint;
 use Drupal\media\Plugin\Validation\Constraint\OEmbedResourceConstraintValidator;
@@ -37,7 +39,7 @@ class OEmbedResourceConstraintValidatorTest extends KernelTestBase {
   /**
    * @covers ::validate
    */
-  public function testValidate() {
+  public function testValidateEmptySource() {
     $media = Media::create([
       'bundle' => $this->createMediaType('oembed:video')->id(),
     ]);
@@ -52,7 +54,55 @@ class OEmbedResourceConstraintValidatorTest extends KernelTestBase {
     $url_resolver = $this->prophesize(UrlResolverInterface::class);
     $url_resolver->getProviderByUrl(Argument::any())->shouldNotBeCalled();
 
-    $value = new class ($media) {
+    $validator = new OEmbedResourceConstraintValidator(
+      $url_resolver->reveal(),
+      $this->container->get('media.oembed.resource_fetcher'),
+      $this->container->get('logger.factory')
+    );
+    $validator->initialize($context->reveal());
+    $validator->validate($this->getValue($media), $constraint);
+  }
+
+  /**
+   * @covers ::validate
+   */
+  public function testValidateUrlResolverInvoked() {
+    $media = Media::create([
+      'bundle' => $this->createMediaType('oembed:video')->id(),
+      'field_media_oembed_video' => 'source value',
+    ]);
+
+    $constraint = new OEmbedResourceConstraint();
+
+    $context = $this->prophesize(ExecutionContextInterface::class);
+
+    $provider = $this->prophesize(Provider::class);
+    $provider->getName()->willReturn('YouTube');
+
+    $url_resolver = $this->prophesize(UrlResolverInterface::class);
+    $url_resolver->getProviderByUrl(Argument::any())->willReturn($provider->reveal());
+    $url_resolver->getResourceUrl(Argument::any())->shouldBeCalledOnce();
+
+    $validator = new OEmbedResourceConstraintValidator(
+      $url_resolver->reveal(),
+      $this->prophesize(ResourceFetcher::class)->reveal(),
+      $this->container->get('logger.factory')
+    );
+    $validator->initialize($context->reveal());
+    $validator->validate($this->getValue($media), $constraint);
+  }
+
+  /**
+   * Wraps a media entity in an anonymous class to mock a field value.
+   *
+   * @param \Drupal\media\Entity\Media $media
+   *   The media object.
+   *
+   * @return object
+   *   The mock field value to validate.
+   */
+  protected function getValue(Media $media) {
+    return new class ($media) {
 
       public function __construct($entity) {
         $this->entity = $entity;
@@ -63,14 +113,6 @@ class OEmbedResourceConstraintValidatorTest extends KernelTestBase {
       }
 
     };
-
-    $validator = new OEmbedResourceConstraintValidator(
-      $url_resolver->reveal(),
-      $this->container->get('media.oembed.resource_fetcher'),
-      $this->container->get('logger.factory')
-    );
-    $validator->initialize($context->reveal());
-    $validator->validate($value, $constraint);
   }
 
 }
