@@ -32,7 +32,6 @@ class Upgrade7Test extends MigrateUpgradeExecuteTestBase {
     'forum',
     'rdf',
     'statistics',
-    'migration_provider_test',
   ];
 
   /**
@@ -87,8 +86,8 @@ class Upgrade7Test extends MigrateUpgradeExecuteTestBase {
       'contact_form' => 3,
       'contact_message' => 0,
       'editor' => 2,
-      'field_config' => 79,
-      'field_storage_config' => 60,
+      'field_config' => 81,
+      'field_storage_config' => 62,
       'file' => 3,
       'filter_format' => 7,
       'image_style' => 6,
