@@ -12,7 +12,7 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\BrowserKit\Client as SymfonyClient;
 use Symfony\Component\Filesystem\Filesystem as SymfonyFilesystem;
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\Lock\Factory;
+use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Lock\Store\FlockStore;
 use Symfony\Component\Process\PhpExecutableFinder;
 use Symfony\Component\Process\Process;
@@ -465,7 +465,7 @@ abstract class BuildTestBase extends TestCase {
    */
   protected function findAvailablePort() {
     $store = new FlockStore(DrupalFilesystem::getOsTemporaryDirectory());
-    $lock_factory = new Factory($store);
+    $lock_factory = new LockFactory($store);
 
     $counter = 100;
     while ($counter--) {
