@@ -1,4 +1,4 @@
-/*!
+;/*!
  * hoverIntent v1.8.1 // 2014.08.11 // jQuery v1.9.1+
  * http://briancherne.github.io/jquery-hoverIntent/
  *
@@ -28,16 +28,14 @@
  * @param  handlerOut  function OR selector for delegation OR undefined
  * @param  selector    selector OR undefined
  * @author Brian Cherne <brian(at)cherne(dot)net>
- */
-
-;(function(factory) {
+ */(function (factory) {
   'use strict';
   if (typeof define === 'function' && define.amd) {
     define(['jquery'], factory);
   } else if (jQuery && !jQuery.fn.hoverIntent) {
     factory(jQuery);
   }
-})(function($) {
+})(function ($) {
   'use strict';
 
   // default configuration values
@@ -54,15 +52,15 @@
   var cX, cY;
 
   // saves the current pointer position coordinates based on the given mousemove event
-  var track = function(ev) {
+  var track = function (ev) {
     cX = ev.pageX;
     cY = ev.pageY;
   };
 
   // compares current and previous mouse positions
-  var compare = function(ev,$el,s,cfg) {
+  var compare = function (ev,$el,s,cfg) {
     // compare mouse positions to see if pointer has slowed enough to trigger `over` function
-    if ( Math.sqrt( (s.pX-cX)*(s.pX-cX) + (s.pY-cY)*(s.pY-cY) ) < cfg.sensitivity ) {
+    if ( Math.sqrt( (s.pX - cX) * (s.pX - cX) + (s.pY - cY) * (s.pY - cY) ) < cfg.sensitivity ) {
       $el.off(s.event,track);
       delete s.timeoutId;
       // set hoverIntent state as active for this element (permits `out` handler to trigger)
@@ -76,17 +74,17 @@
       // set previous coordinates for next comparison
       s.pX = cX; s.pY = cY;
       // use self-calling timeout, guarantees intervals are spaced out properly (avoids JavaScript timer bugs)
-      s.timeoutId = setTimeout( function(){compare(ev, $el, s, cfg);} , cfg.interval );
+      s.timeoutId = setTimeout( function () {compare(ev, $el, s, cfg);} , cfg.interval );
     }
   };
 
   // triggers given `out` function at configured `timeout` after a mouseleave and clears state
-  var delay = function(ev,$el,s,out) {
+  var delay = function (ev,$el,s,out) {
     delete $el.data('hoverIntent')[s.id];
     return out.apply($el[0],[ev]);
   };
 
-  $.fn.hoverIntent = function(handlerIn,handlerOut,selector) {
+  $.fn.hoverIntent = function (handlerIn,handlerOut,selector) {
     // instance ID, used as a key to store and retrieve state information on an element
     var instanceId = INSTANCE_COUNT++;
 
@@ -104,7 +102,7 @@
     }
 
     // A private function for handling mouse 'hovering'
-    var handleHover = function(e) {
+    var handleHover = function (e) {
       // cloned event to pass to handlers (copy required for event object to be passed in IE)
       var ev = $.extend({},e);
 
@@ -130,7 +128,7 @@
       if (state.timeoutId) { state.timeoutId = clearTimeout(state.timeoutId); }
 
       // namespaced event used to register and unregister mousemove tracking
-      var mousemove = state.event = 'mousemove.hoverIntent.hoverIntent'+instanceId;
+      var mousemove = state.event = 'mousemove.hoverIntent.hoverIntent' + instanceId;
 
       // handle the event, based on its type
       if (e.type === 'mouseenter') {
@@ -141,14 +139,14 @@
         // update "current" X and Y position based on mousemove
         $el.off(mousemove,track).on(mousemove,track);
         // start polling interval (self-calling timeout) to compare mouse coordinates over time
-        state.timeoutId = setTimeout( function(){compare(ev,$el,state,cfg);} , cfg.interval );
+        state.timeoutId = setTimeout( function () {compare(ev,$el,state,cfg);} , cfg.interval );
       } else { // "mouseleave"
         // do nothing if not already active
         if (!state.isActive) { return; }
         // unbind expensive mousemove event
         $el.off(mousemove,track);
         // if hoverIntent state is true, then call the mouseOut function after the specified delay
-        state.timeoutId = setTimeout( function(){delay(ev,$el,state,cfg.out);} , cfg.timeout );
+        state.timeoutId = setTimeout( function () {delay(ev,$el,state,cfg.out);} , cfg.timeout );
       }
     };
 
